API

This page contains documentation of the public API of MRINavigator. In the Julia REPL one can access this documentation by entering the help mode with ? and then writing the function for which the documentation should be shown.

Run compact pipeline

Missing docstring.

Missing docstring for MRINavigator.defaultNavParams :: Tuple{}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.runNavPipeline :: Dict{Symbol, Any}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.saveNoise :: Tuple{String, String}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.loadRawData :: Tuple{Dict{Symbol, Any}}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.convertRawToAcq :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

<!–

Coil sensitivity maps

MRINavigator.CompSensit — Function

sensit = CompSensit(acq::AcquisitionData, thresh = 0.13)

Compute the coils sensitivity maps with masking tuned for spinal cord imaging. Use MRICoilSensitivities.jl from MRIReco.jl alternatively.

MRIReco reference: https://onlinelibrary.wiley.com/doi/epdf/10.1002/mrm.28792

Arguments

acqData::RawAcquisitionData - acquisition data structure obtained converting raw data with MRIReco.jl

tresh::Float64 - masking treshold: increase for reduced mask size, decrease for extended mask size

MRINavigator.CompRoughMask — Function

mask = CompRoughMask(acq::AcquisitionData, slices::Int64, thresh)

Return a rough mask for multiple slices that may not be homogeneous.

Arguments

acqData::RawAcquisitionData - acquisition data structure obtained converting raw data with MRIReco.jl

slices::Int64 - number of slices in acquisition data

tresh::Float64 - masking treshold: increase for reduced mask size, decrease for extended mask size

MRIReco reference: https://onlinelibrary.wiley.com/doi/epdf/10.1002/mrm.28792

MRINavigator.ResizeSensit! — Function

sensit = ResizeSensit!(sensit::Array{Complex{T},4}, acqMap::AcquisitionData, acqData::AcquisitionData)

Resize and resample the coil sensitivity map to match the acquisition data field of view and resolution. This step is needed for the image reconstruction to run. Image data and reference data must have the same slice center.

Arguments

sensit::Array{Complex{T},4} - output of CompSensit(acq::AcquisitionData, thresh)

acqMap::RawAcquisitionData - acquisition data structure obtained converting raw reference data with MRIReco.jl

acqData::RawAcquisitionData - acquisition data structure obtained converting raw data with MRIReco.jl

MRIReco reference: https://onlinelibrary.wiley.com/doi/epdf/10.1002/mrm.28792

MRINavigator.CompResizeSaveSensit — Function

CompResizeSaveSensit(acqMap::AcquisitionData, acqData::AcquisitionData, path_sensit::String)

Compute, resize to the image data dimension and save the coils sensitivity maps with masking tuned for spinal cord imaging. Use MRICoilSensitivities.jl from MRIReco.jl alternatively.

MRIReco reference: https://onlinelibrary.wiley.com/doi/epdf/10.1002/mrm.28792

Arguments

acqMap::RawAcquisitionData - acquisition data structure obtained converting raw data with MRIReco.jl

acqData::RawAcquisitionData - acquisition data structure obtained converting raw data with MRIReco.jl

tresh::Float64 - masking treshold: increase for reduced mask size, decrease for extended mask size

Find centerline

Missing docstring.

Missing docstring for MRINavigator.findCenterline :: Tuple{Dict{Symbol, Any}}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.ReconstructMap :: Tuple{String}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.ReconstructSaveMap :: Tuple{String, String}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.callSCT :: Tuple{Dict{Symbol, Any}}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.comp_centerline_pos :: Tuple{additionalNavInput}. Check Documenter's build log for details.

Utils

Missing docstring.

Missing docstring for MRINavigator.Reconstruct :: AcquisitionData, Array{Complex{T},4}, Union{Array{Complex{T}},Nothing} where {T}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.directreco :: Tuple{MRIBase.AcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.niftiSaveImg :: Union{Tuple{T}, Tuple{AbstractArray{T}, MRIBase.AcquisitionData, String}} where T. Check Documenter's build log for details.

Navigator correction

Missing docstring.

Missing docstring for MRINavigator.NavCorr! :: Array{Complex{T}, 4}, AcquisitionData, Dict{Symbol, Any}, additionalNavInput where{T}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.wrap_corr! :: Tuple{Array{Float64, 4}, Matrix{Int8}, VecOrMat{Float64}, Int64}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.find_wrapped :: Tuple{Array{Float64, 4}, Matrix{Float64}, Matrix{Float64}, Int64}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.TE_corr! :: Union{Tuple{T}, Tuple{Array{T, 4}, MRIBase.AcquisitionData, Float64, Float64, Int64, Int64}} where T. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.apply_corr! :: Union{Tuple{T}, Tuple{Array{T, 4}, MRIBase.AcquisitionData, Int64, Int64, Int64, Int64}} where T. Check Documenter's build log for details.

Adjust data

Missing docstring.

Missing docstring for MRINavigator.OrderSlices! :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.ExtractFlags :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.ExtractNoiseData! :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.ReverseBipolar! :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.RemoveRef! :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.CopyTE! :: Tuple{MRIBase.RawAcquisitionData, MRIBase.AcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.AdjustSubsampleIndices! :: Tuple{MRIBase.AcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.ExtractNavigator :: Tuple{MRIBase.RawAcquisitionData}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.selectEcho! :: Tuple{MRIBase.AcquisitionData, Vector{Int64}}. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.selectSlice! :: Union{Tuple{T}, Tuple{MRIBase.AcquisitionData, Vector{Int64}}, Tuple{MRIBase.AcquisitionData, Vector{Int64}, Union{Nothing, Array{Complex{T}, 4}}}, Tuple{MRIBase.AcquisitionData, Vector{Int64}, Union{Nothing, Array{Complex{T}, 4}}, Union{Nothing, Matrix{Float64}}}} where T. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MRINavigator.additionalNavInput :: Union{Tuple{Matrix{ComplexF32}, MRIBase.RawAcquisitionData, MRIBase.AcquisitionData}, Tuple{Matrix{ComplexF32}, MRIBase.RawAcquisitionData, MRIBase.AcquisitionData, Union{Nothing, MRIBase.AcquisitionData}}, Tuple{Matrix{ComplexF32}, MRIBase.RawAcquisitionData, MRIBase.AcquisitionData, Union{Nothing, MRIBase.AcquisitionData}, Union{Nothing, Matrix{Float64}}}, Tuple{Matrix{ComplexF32}, MRIBase.RawAcquisitionData, MRIBase.AcquisitionData, Union{Nothing, MRIBase.AcquisitionData}, Union{Nothing, Matrix{Float64}}, Union{Nothing, Matrix{Float64}}}, Tuple{Matrix{ComplexF32}, MRIBase.RawAcquisitionData, MRIBase.AcquisitionData, Union{Nothing, MRIBase.AcquisitionData}, Union{Nothing, Matrix{Float64}}, Union{Nothing, Matrix{Float64}}, Union{Nothing, Vector{Float64}}}}. Check Documenter's build log for details.

–>